requirements.txt
deepagents>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain>=0.1.0
python-dotenv>=1.0.0
